Add r8168 kernel module package to build
The make-kpkg and module-assistant facilities for creating kernel
modules packages don't create the modules from a Debian source package
in the typical way.  It's easy enough to use these normally, but in
the chroot we need to set up quite a bit of apparatus to make it work
properly.

After the kernel modules are built:
- Rebuild the PPA with kernel pkgs and update the pbuilder chroot
- Clone the 'kernel-modules-deb' git repo
- Run the chroot in 'execute' mode to build the kmodules:
  - Pass in the PPA with kernel pkgs
  - Pass in a tarball of kernel-modules-deb
  - Run the pbuild/build-kmodules.sh script
    - Install pkgs required for build
    - Build modules in 'kernel-modules-deb'
  - Pass the built modules back out by bind-mounting the 'pkgs' directory