Set up linux kernel source package in a chroot
The kernel source package needs to be configured before it can be
created with dpkg-source.  During this step, the i-pipe patch is taken
from the linux-patch-xenomai package and added as a feature set patch.

The linux-patch-xenomai package is generated in this same system, and
can't be expected to already be installed on the host.  Instead,
configure the source package from within a random chroot.