Fix linux package list
The linux package is particularly annoying, since it embeds the
featureset and 'flavour' names in the package name.

The problem is that unlike other packages, the names differ between
each build-arch.

This patch adds a map to the Xenomai rules, rules to compute the names
in the Linux rules, and rules to pull in arch-specific names from the
main Makefile.