Add `linux-latest` package
The `linux-latest` package provides a way to drag in kernel packages
without needing to specify the ABI name, which is embedded in the
kernel packages' names.

The `git/linux-latest-deb` submodule was cloned from the upstream
Debian package and modified for these kernels.

A `-p PACKAGE` argument is added to
`pbuild/linux-unpacked-chroot-script.sh` so that it can be used for
more than just the `linux` package, since the `linux-latest` source
package configuration depends on the `linux-support` package.

The `linux-support` package build, previously disabled, is re-added.